Move the MaybeTls enum and impls into a separate source
